requirements.txt
Flask==2.2.3
numpy==1.23.4
psycopg2==2.9.6
python-dotenv==1.0.0
